renames PreviewTable to BudgetLinesTable
