--- frontend/src/components/UI/PreviewTable/index.js
@@ -1,1 +1,0 @@
-export { default } from "./PreviewTable";
